Initial commit to the project
This project is a simple way for me to learn Go, so I thought I'd start
with something that makes sense to me, loans and finance. Given that,
this package is aimed at providing a series of functions that help with
preparing for taking out a loan or for adjusting an existing one.