Better structure repayment period to number of repayments
The switch statement that is replaced in this commit just didn't sit
right with me. That said, to have to create an in-situ constructor to
initialise the map doesn't sit totally well with me either. However,
using a string/int map is, to me, far clearer and cleaner than a switch.